v4.11
Version 4.11: 2024-12-30

    new: add planning tool to find when satellite is up for both DE and DX
    new: add watch list filters for sub-band modes and against ADIF file
    new: add support for N1MM and DXLog UDP DE-spotted packets
    new: add option to rotate between showing call sign and title
    new: add options to sort ADIF list by band or call
    change: DX Cluster pane only lists newest spot on a given band
    change: enforce upper case call sign
    fix: clicking within tall left pane could still trigger sat name on map
    fix: better timezone retrying after network error
    fix: possible fix for getNVMaidenhead invalid 79

+++ ESPHamClock/string.cpp
@@ -255,7 +255,7 @@ char *formatAge (time_t age, char *line, int line_l, int cols)
 {
     // eh?
     if (age < 0) {
-        Serial.printf ("formatAge(%ld,%d) resetting negative age to zero\n", age, cols);
+        Serial.printf ("formatAge(%ld,%d) resetting negative age to zero\n", (long)age, cols);
         age = 0;
     }
